check for a valid objectId

+++ routes/apartments.js
@@ -11,6 +11,7 @@ router.get('/', async (req, res) => {
 
 // get an apartment
 router.get('/:id', async (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
   const apartment = await Apartment.findById(req.params.id);
   if (!apartment) return res.status(404).send('apartment not found');
   res.status(200).send(apartment);
@@ -44,6 +45,7 @@ router.put('/:id', async (req, res) => {
 
   const { value, error } = validateApartment(apartment);
   if (error) return res.status(500).send(error.details[0].message);
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
 
   apartment = await Apartment.findByIdAndUpdate(req.params.id, value, { new: true });
   if (!apartment) return res.status(404).send('apartment not found');
@@ -52,6 +54,7 @@ router.put('/:id', async (req, res) => {
 
 // delete an apartment
 router.delete('/:id', async (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
   const apartment = await Apartment.findByIdAndRemove(req.params.id);
   if (!apartment) return res.status(200).send('apartment not found');
   res.status(200).send(apartment);
