+++ routes/houseHelps.js
@@ -10,6 +10,7 @@ router.get('/', async (req, res) => {
 
 // return a particular house help
 router.get('/:id', async (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
   const houseHelp = await HouseHelp.findById(req.params.id);
   if (!houseHelp) return res.status(404).send('house help not found');
   res.status(200).send(houseHelp);
@@ -43,6 +44,7 @@ router.put('/:id', async (req, res) => {
 
   const { value, error } = validateHouseHelp(houseHelp);
   if (error) return res.status(500).send(error.details[0].message);
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
 
   houseHelp = await HouseHelp.findByIdAndUpdate(req.params.id, value, {
     new: true,
@@ -53,6 +55,7 @@ router.put('/:id', async (req, res) => {
 
 // delete house help
 router.delete('/:id', async (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
   const houseHelp = await HouseHelp.findByIdAndDelete(req.params.id);
   if (!houseHelp) return res.status(404).send('house help not found');
   res.status(200).send(houseHelp);
