+++ routes/visitors.js
@@ -11,6 +11,7 @@ router.get('/', async (req, res) => {
 
 // get a visitor
 router.get('/:id', async (req, res, next) => {
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
   const visitor = await Visitor.findById(req.params.id);
   if (!visitor) return res.status(404).send('incorrect ID');
   res.status(200).send(visitor);
@@ -49,6 +50,7 @@ router.put('/:id', async (req, res, next) => {
 
   const { value, error } = validateVisitor(visitor);
   if (error) return res.status(500).send(error.details[0].message);
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
 
   visitor = await Visitor.findByIdAndUpdate(req.params.id, value, {
     new: true,
@@ -59,6 +61,7 @@ router.put('/:id', async (req, res, next) => {
 
 // delete a visitor
 router.delete('/:id', async (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid object id');
   const visitor = await Visitor.findByIdAndRemove(req.params.id);
   if (!visitor) return res.status(404).send('visitor not found');
   res.status(200).send(visitor);
